refactor: Changed position packaging of utils
