+++ app/src/main/groovy/org/cpuInfoFetcher/AMDSpecificationsFetcher.groovy
@@ -1,5 +1,7 @@
 package org.cpuinfofetcher
 
+import org.cpuinfofetcher.utils.InteractiveHTMLScraper
+
 import java.nio.file.Path
 import java.nio.file.Paths
 import java.nio.file.Files
