+++ app/src/main/groovy/org/cpuInfoFetcher/AmperaSpecificationsFetcher.groovy
@@ -1,5 +1,7 @@
 package org.cpuinfofetcher
 
+import org.cpuinfofetcher.utils.HTMLScraper
+
 import java.nio.file.Path
 import java.nio.file.Paths
 import java.nio.file.Files
